Add files via upload
The data set used in this paper is StudentsPerformance.csv, which contains the performance of 
1000 students in Mathematics, Reading, and Writing. These categorical variables include gender, 
race/ethnicity, the parental level of education, the type of lunch students are entitled to, and 
whether the students completed a test preparation course. This data was collected to analyse the 
effect of different socio-demographic factors on students’ performance. Performance levels for 
three other subjects are represented in numbers from 0 to 100; quantitative values of learning 
outcomes are presented. This open-source dataset is commonly used in educational research to 
study the determinants of learners’ performance and explore the policy strategies necessary to 
improve learners’ performance (Zagatti et al., 2022). It becomes suitable to use clustering 
methodologies to define the internal structure of the clusters based on values and demographic 
data that may help in educational policies and strategies for optimised learning for students.